adding Platform variable to more stylesheets, adding ProfileScreenStyles stylesheet
